# docs/swagger.yaml
basePath: /
definitions:
...
  helpers.SwaggerError400Response:
    properties:
      message:
        example: Input tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError401Response:
    properties:
      message:
        example: Token tidak ditemukan atau tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError403Response:
    properties:
      code:
        example: ERR_PERMISSION_DENIED
        type: string
      message:
        example: 'Akses ditolak: Anda tidak memiliki akses untuk fitur ini'
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError404Response:
    properties:
      message:
        example: Data tidak ditemukan
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError500Response:
    properties:
      message:
        example: Terjadi kesalahan internal pada server
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerSuccessResponse:
    properties:
      data: {}
      message:
        example: Berhasil memproses data
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.Guru:
    properties:
      alamat:
        description: Deprecated Database Columns (Retained for Non-Destructive Migrations)
        example: Jl. Diponegoro No. 25
        type: string
      alamat_detail:
        example: Jl. Diponegoro No. 25
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      desa:
        example: Sidorejo Lor
        type: string
      gelar:
        example: S.Pd.
        type: string
      id:
        example: 1
        type: integer
      jenis_kelamin:
        example: Laki-laki
        type: string
      kabupaten:
        example: Salatiga
        type: string
      kecamatan:
        example: Sidorejo
        type: string
      mapel_ids:
        description: Virtual field for frontend multiselect checkboxes
        example:
        - 1
        - 2
        items:
          type: integer
        type: array
      nama:
        example: Ahmad Guru
        type: string
      nip:
        example: GR001
        type: string
      no_hp:
        example: "081234567890"
        type: string
      photo_url:
        example: /uploads/profile/guru_1.jpg
        type: string
      provinsi:
        description: Address Normalization Fields
        example: Jawa Tengah
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        example: 3
        type: integer
    type: object
...
  model.Jadwal:
    properties:
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      guru:
        $ref: '#/definitions/model.Guru'
      guru_id:
        example: 1
        type: integer
      hari:
        example: Senin
        type: string
      id:
        example: 1
        type: integer
      jam_mulai:
        example: "07:00"
        type: string
      jam_selesai:
        example: "09:00"
        type: string
      kelas:
        $ref: '#/definitions/model.Kelas'
      kelas_id:
        example: 1
        type: integer
      mapel:
        $ref: '#/definitions/model.Mapel'
      mapel_id:
        example: 1
        type: integer
      semester:
        example: Ganjil
        type: string
      tahun_ajaran:
        example: 2026/2027
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
    type: object
  model.JadwalResponseList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Jadwal'
        type: array
      message:
        example: Berhasil mengambil data jadwal pelajaran
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.JadwalResponseSingle:
    properties:
      data:
        $ref: '#/definitions/model.Jadwal'
      message:
        example: Berhasil mengambil detail jadwal pelajaran
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.Kelas:
    properties:
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      jurusan:
        example: Rekayasa Perangkat Lunak
        type: string
      kapasitas:
        example: 36
        type: integer
      nama_kelas:
        example: X RPL 1
        type: string
      tingkat:
        example: X
        type: string
      total_siswa:
        example: 32
        type: integer
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      wali_kelas:
        $ref: '#/definitions/model.Guru'
      wali_kelas_id:
        example: 1
        type: integer
    type: object
  model.KelasResponseList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Kelas'
        type: array
      message:
        example: Berhasil mengambil data kelas
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.KelasResponseSingle:
    properties:
      data:
        $ref: '#/definitions/model.Kelas'
      message:
        example: Berhasil mengambil detail kelas
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.Mapel:
    properties:
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      is_umum:
        example: true
        type: boolean
      jam:
        example: 4
        type: integer
      jurusan:
        example: Umum
        type: string
      kode_mapel:
        example: MP001
        type: string
      nama_mapel:
        example: Matematika
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
    type: object
...
  model.User:
    properties:
      alamat:
        example: Jl. Diponegoro No. 25, Salatiga
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      email:
        example: ahmad.guru@sekolah.com
        type: string
      gelar:
        example: S.Pd.
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      is_first_login:
        example: true
        type: boolean
      jenis_kelamin:
        example: Laki-laki
        type: string
      last_login_at:
        type: string
      nama:
        example: Ahmad Guru
        type: string
      nip:
        description: Ignored by GORM, populated for unified API login response
        example: GR001
        type: string
      nis:
        example: "12345678"
        type: string
      no_hp:
        example: "081234567890"
        type: string
      role:
        example: guru
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: praktikum@example.com
    name: Praktikum Pemrograman III
  description: Dokumentasi API backend Sistem Informasi Sekolah menggunakan Golang
    Fiber, GORM, PostgreSQL, dan JWT.
  title: API Sistem Informasi Sekolah
  version: "1.0"
paths:
...
  /api/jadwal/:
    get:
      consumes:
      - application/json
      description: Mengambil seluruh data jadwal pelajaran dengan berbagai filter
        opsional.
      parameters:
      - description: Filter ID Kelas
        in: query
        name: kelas_id
        type: string
      - description: Filter ID Mapel
        in: query
        name: mapel_id
        type: string
      - description: Filter ID Guru
        in: query
        name: guru_id
        type: string
      - description: Filter Hari
        in: query
        name: hari
        type: string
      - description: Filter Tahun Ajaran (e.g. 2026/2027)
        in: query
        name: tahun_ajaran
        type: string
      - description: Filter Semester (e.g. Ganjil/Genap)
        in: query
        name: semester
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JadwalResponseList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ambil semua data jadwal
      tags:
      - Jadwal
    post:
      consumes:
      - application/json
      description: Menambahkan data jadwal pelajaran baru.
      parameters:
      - description: Payload data jadwal
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Jadwal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.JadwalResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Tambah data jadwal
      tags:
      - Jadwal
  /api/jadwal/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data jadwal pelajaran berdasarkan ID. Hanya dapat diakses
        oleh Admin.
      parameters:
      - description: ID Jadwal
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.SwaggerSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.SwaggerError403Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Hapus data jadwal
      tags:
      - Jadwal
    get:
      consumes:
      - application/json
      description: Mengambil detail data satu jadwal pelajaran berdasarkan ID.
      parameters:
      - description: ID Jadwal
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JadwalResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ambil data jadwal berdasarkan ID
      tags:
      - Jadwal
    put:
      consumes:
      - application/json
      description: Mengubah data jadwal pelajaran berdasarkan ID.
      parameters:
      - description: ID Jadwal
        in: path
        name: id
        required: true
        type: integer
      - description: Payload data jadwal
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Jadwal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JadwalResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ubah data jadwal
      tags:
      - Jadwal
  /api/kelas/:
    get:
      consumes:
      - application/json
      description: Mengambil seluruh data kelas. Mendukung pencarian nama kelas dan
        filter tingkat.
      parameters:
      - description: Cari berdasarkan nama kelas
        in: query
        name: search
        type: string
      - description: Filter tingkat kelas (e.g. X, XI, XII)
        in: query
        name: tingkat
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.KelasResponseList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ambil semua data kelas
      tags:
      - Kelas
    post:
      consumes:
      - application/json
      description: Menambahkan data kelas baru.
      parameters:
      - description: Payload data kelas
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Kelas'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.KelasResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Tambah data kelas
      tags:
      - Kelas
  /api/kelas/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data kelas berdasarkan ID. Hanya dapat diakses oleh Admin.
      parameters:
      - description: ID Kelas
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.SwaggerSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.SwaggerError403Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Hapus data kelas
      tags:
      - Kelas
    get:
      consumes:
      - application/json
      description: Mengambil detail data satu kelas berdasarkan ID.
      parameters:
      - description: ID Kelas
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.KelasResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ambil data kelas berdasarkan ID
      tags:
      - Kelas
    put:
      consumes:
      - application/json
      description: Mengubah data kelas berdasarkan ID.
      parameters:
      - description: ID Kelas
        in: path
        name: id
        required: true
        type: integer
      - description: Payload data kelas
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Kelas'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.KelasResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ubah data kelas
      tags:
      - Kelas
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"